+++ integration_tests/suite/test_calls.py
@@ -3332,6 +3332,25 @@ class TestCallSendDTMF(RealAsteriskIntegrationTest):
 
         until.assert_(amid_dtmf_events_received, tries=10)
 
+    def test_put_dtmf_tenant_isolation(self):
+        user_uuid_1 = str(uuid.uuid4())
+        tenant_uuid_1 = str(uuid.uuid4())
+        tenant_uuid_2 = str(uuid.uuid4())
+        channel_id_1 = self.given_call_not_stasis(tenant_uuid=tenant_uuid_1)
+        channel_id_2 = self.given_call_not_stasis(tenant_uuid=tenant_uuid_2)
+        user_calld = self.make_user_calld(user_uuid_1, tenant_uuid=tenant_uuid_1)
+        dtmf = '12*#'
+
+        # send dtmf to channel from other tenant = NOK
+        with pytest.raises(CalldError) as exc_info:
+            user_calld.calls.send_dtmf_digits(channel_id_2, dtmf)
+
+        calld_error = exc_info.value
+        assert calld_error.status_code == 404, calld_error
+
+        # send dtmf to channel from same tenant = OK
+        user_calld.calls.send_dtmf_digits(channel_id_1, dtmf)
+
     def test_put_dtmf_from_user(self):
         user_uuid = str(uuid.uuid4())
         token = self.given_user_token(user_uuid)
@@ -3385,8 +3404,9 @@ class TestCallSendDTMF(RealAsteriskIntegrationTest):
 
         until.assert_(amid_dtmf_events_received, tries=10)
 
-    def given_call_not_stasis(self, user_uuid=None):
+    def given_call_not_stasis(self, user_uuid=None, tenant_uuid=None):
         user_uuid = user_uuid or str(uuid.uuid4())
+        tenant_uuid = tenant_uuid or VALID_TENANT
         call = self.ari.channels.originate(
             endpoint=ENDPOINT_AUTOANSWER,
             context='local',
@@ -3394,7 +3414,7 @@ class TestCallSendDTMF(RealAsteriskIntegrationTest):
             variables={
                 'variables': {
                     'WAZO_USERUUID': user_uuid,
-                    '__WAZO_TENANT_UUID': VALID_TENANT,
+                    '__WAZO_TENANT_UUID': tenant_uuid,
                 }
             },
         )